requirements.txt
asgiref==3.3.4
django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
